Progress.

+++ src/main/java/piotrrr/ga/schema/Position.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class Position {
-    private long x;
-    private long y;
-    private long z;
+  private int x;
+  private int y;
+  private int z;
 }
